New Development: No             JIRA Issue: cas-3230         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: No              Module(s): Module Names change impacts.            Description: Put explicit typecasts in front of all floating-point constants used for internal thresholds. this is an attempt to reconcile slight differences between unit-test flag counts on 32 vs 64 bit.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@18607 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/flagging/implement/Flagging/FlagAgentTimeFreqCrop.cc
@@ -373,7 +373,7 @@ void FlagAgentTimeFreqCrop :: fitBaseAndFlag(String fittype, String direction, V
 		}
 	      
 	      // Stop iterating if the deviation of the normalized data from the mean is less than 10%
-	      if(fabs(temp-sd) < 0.1)break;
+	      if(fabs(temp-sd) < (Double)0.1)break;
 	      // else go on for 5 iterations
 	      temp=sd;
 	    }//for loop
@@ -478,7 +478,7 @@ Float FlagAgentTimeFreqCrop :: calcVar(Vector<Float> &vect, Vector<Bool> &flag, 
     {
       if(flag[i]==False)
 	{
-	  validvals[cnt] = fit[i] < 1e-6 ? 0.0 : fabs( (vect[i] - fit[i])/fit[i] );
+	  validvals[cnt] = fit[i] < (Double)1e-6 ? (Double)0.0 : fabs( (vect[i] - fit[i])/fit[i] );
 	  cnt++;
 	}
     }
@@ -690,7 +690,6 @@ void FlagAgentTimeFreqCrop :: polyFit(Vector<Float> &data,Vector<Bool> &flag, Ve
   
   LinearFit<Double> fitter;
   Polynomial<Double> combination(deg);
-  
   
   combination.setCoefficient(0,0.0);
   if (deg >= 1) combination.setCoefficient(1, 0.0);
